+++ lib/bones_api_server.dart
@@ -2,6 +2,6 @@
 library bones_api_server;
 
 export 'package:bones_api/bones_api.dart';
+export 'package:dart_spawner/dart_spawner.dart';
 
-export 'src/bones_api_isolate.dart';
 export 'src/bones_api_server.dart';